Add metrics UnAuthUserGotZero & <user>GotNonZero
UnAuthUserGotZero is, out of all the requests that were zero,
the percentage of them where the user was None/unauthenticated.
This will tell us if a large portion of the badge returning
zero requests are due to unauthenticated users.

<user>GotNonZero is a count or the number of requests that the
given user got a badge count that wasn't zero. This will tell us
if there are a small portion of users for which the badge count
is consistently reporting non zero.